Fix emptyTrackTest and minor improvements to trackOptionstest
