+++ app/src/androidTest/java/net/osmtracker/activity/TrackManagerTest.java
@@ -1,24 +1,23 @@
 package net.osmtracker.activity;
 
-import android.app.Activity;
-import android.support.test.InstrumentationRegistry;
 import android.support.test.espresso.Espresso;
 import android.support.test.espresso.assertion.ViewAssertions;
-import android.support.test.espresso.core.internal.deps.guava.collect.Iterables;
 import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.lifecycle.ActivityLifecycleMonitorRegistry;
-import android.support.test.runner.lifecycle.Stage;
+import android.view.View;
+import android.widget.TextView;
 
 import net.osmtracker.R;
 import net.osmtracker.util.TestUtils;
 
+import static android.support.test.espresso.matcher.ViewMatchers.hasSibling;
+import static org.hamcrest.Matchers.allOf;
+import static org.hamcrest.Matchers.not;
 import org.junit.Rule;
 import org.junit.Test;
 
-import java.util.Collection;
-import java.util.Iterator;
-
+import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.Espresso.openActionBarOverflowOrOptionsMenu;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.longClick;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
@@ -34,10 +33,37 @@ public class TrackManagerTest {
     @Test
     public void emptyTracksTest() {
 
-        //FIXME: Only works if the db is empty.
+        TextView tv_emptymsg = (TextView) mRule.getActivity().findViewById(R.id.trackmgr_empty);
+        if (tv_emptymsg.getVisibility() == View.VISIBLE) {
+            onView(withText(TestUtils.getStringResource(R.string.trackmgr_empty)))
+                    .check(ViewAssertions.matches(isDisplayed()));
 
+            // TODO: check that settings menu only shows settings and about message.
+            //openActionBarOverflowOrOptionsMenu(getInstrumentation().getTargetContext());
+
+            // create a new track
+            onView(withId(R.id.trackmgr_hint_icon)).perform(click());
+            checkToastIsShownWith(TestUtils.getStringResource(R.string.tracklogger_waiting_gps));
+            // stop and save
+            onView(withId(R.id.tracklogger_menu_stoptracking)).perform(click());
+
+        }
+
+        // There is at least one track
+        onView(withText(TestUtils.getStringResource(R.string.trackmgr_empty)))
+                .check(ViewAssertions.matches(not(isDisplayed())));
+
+        // delete all tracks
+        openActionBarOverflowOrOptionsMenu(getInstrumentation().getTargetContext());
+        onView(withText(TestUtils.getStringResource(R.string.menu_deletetracks))).perform(click());
+        //confirmation button
+        onView(withText(TestUtils.getStringResource(R.string.menu_deletetracks))).perform(click());
+
+        // check empty tracks list message
         onView(withText(TestUtils.getStringResource(R.string.trackmgr_empty)))
                 .check(ViewAssertions.matches(isDisplayed()));
+
+
     }
 
     @Test
@@ -45,9 +71,10 @@ public class TrackManagerTest {
 
         // create a new track
         onView(withId(R.id.trackmgr_hint_icon)).perform(click());
+        checkToastIsShownWith(TestUtils.getStringResource(R.string.tracklogger_waiting_gps));
 
         // get Title from track Logger (Actual activity)
-        String trackLoggerActivityTitle = getActivityInstance().getTitle().toString();
+        String trackLoggerActivityTitle = TestUtils.getActivityInstance().getTitle().toString();
         String baseTitleText = TestUtils.getStringResource(R.string.tracklogger) +  ": #";
         String strTrackId = trackLoggerActivityTitle.replace(baseTitleText, "");
 
@@ -60,14 +87,18 @@ public class TrackManagerTest {
         //resume and stop tracking
         onView(withText("#"+strTrackId)).perform(longClick());
         onView(withText(TestUtils.getStringResource(R.string.trackmgr_contextmenu_resume))).perform(click());
+        checkToastIsShownWith(TestUtils.getStringResource(R.string.tracklogger_waiting_gps));
+
         Espresso.pressBack();
         onView(withText("#"+strTrackId)).perform(longClick());
         onView(withText(TestUtils.getStringResource(R.string.trackmgr_contextmenu_stop))).perform(click());
 
-        // TODO: check GPX contents
+        // export track.  TODO: check GPX contents
         onView(withText("#"+strTrackId)).perform(longClick());
         onView(withText(TestUtils.getStringResource(R.string.trackmgr_contextmenu_export))).perform(click());
-        checkToastIsShownWith(TestUtils.getStringResource(R.string.various_export_finished));
+        onView(allOf(withId(R.id.trackmgr_item_statusicon), hasSibling(withText("#"+strTrackId))))
+                .check(ViewAssertions.matches(isDisplayed()));
+
 
         // delete track
         onView(withText("#"+strTrackId)).perform(longClick());
@@ -75,19 +106,5 @@ public class TrackManagerTest {
         onView(withText("#"+strTrackId)).check(ViewAssertions.doesNotExist());
     }
 
-    //https://stackoverflow.com/questions/38737127/espresso-how-to-get-current-activity-to-test-fragments?noredirect=1&lq=1
-    private Activity getActivityInstance(){
-        final Activity[] currentActivity = {null};
-
-        InstrumentationRegistry.getInstrumentation().runOnMainSync(new Runnable(){
-            public void run(){
-                Collection<Activity> resumedActivity = ActivityLifecycleMonitorRegistry.getInstance().getActivitiesInStage(Stage.RESUMED);
-                Iterator<Activity> it = resumedActivity.iterator();
-                currentActivity[0] = it.next();
-            }
-        });
-
-        return currentActivity[0];
-    }
 
 }
